# cabana: optimize get_raw_value() function for CAN signal extraction