Improves the performance of the get_raw_value() function in DBC signal handling.

### Reason for optimization:
Cabana must efficiently parse all CAN messages in real time and rapidly process up to 30 minutes of CAN data for charting. Optimizing signal extraction directly enhances responsiveness and the overall user experience.

### Changes:
1. Added a fast path for signals fits in a single byte (the most common case), 
2. Simplified multi-byte signal extraction logic
3. Reduced the number of bitwise operations
4. Added early bounds checking
5. Changed `int64_t val` to `uint64_t val` for safe bitwise operations

### Testing:
The updated function was tested and produces identical results to the original implementation for:
- Single-byte signals (signed and unsigned)
- Multi-byte signals spanning byte boundaries
- Little and big endian signals
- Edge cases (out of bounds, sign extension)

### Performance:

| Version | Single-byte | Multi-byte |
|---------|-------------|------------|
| Master  | 12 ns/call  | 17 ns/call |
| PR      |  4 ns/call  |  11 ns/call |

This PR significantly improves extraction speed for both single- and multi-byte signals, while maintaining correctness